Synthetic graph tuning (#117)
* Updated location graph translation to use visits_per_day ~ poisson(lambda=4.61) and the visit times to a resolution from hours to seconds

* feat: switched to letting the number of average visits per day be a floating point number rather than just an int

* fix: liniter errors in generate and translation strategies

+++ src/readers/Parse.cpp
@@ -53,7 +53,7 @@ void parse(int argc, char **argv, Arguments *args) {
     }
 
     // Edge degree.
-    onTheFly->averageVisitsPerDay = atoi(argv[++argNum]);
+    onTheFly->averageVisitsPerDay = atof(argv[++argNum]);
 
     // Chare data
     onTheFly->locationPartitionGrid.width = atoi(argv[++argNum]);
